% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLittleBreakAddin.R
\name{insertLittleBreakAddin}
\alias{insertLittleBreakAddin}
\title{Insert little section break.}
\usage{
insertLittleBreakAddin()
}
\description{
Call this function as an addin to insert section breaks that I used regularly
in my code. Meant to save some typing.
}
